// src/types/processes.ts
export interface ProcessTotalApiItem {
  count: number
  failed: number
}

export interface ProcessTotalApiResponse {
  filter: unknown[]
  items: ProcessTotalApiItem[]
  paging: {
    itemsPerPage: number
    lastPage: number
    nextPage: number
    page: number
    previousPage: number
    total: number
  }
  search: string | null
  sorter: Array<{ column: string; direction: string }>
}

// API Response Types for Graph Data
export interface ProcessGraphApiItem {
  created: string
  topologyId: string
  success: number
  failed: number
}

export interface ProcessGraphApiResponse {
  filter: unknown[]
  items: ProcessGraphApiItem[]
  paging: {
    itemsPerPage: number
    lastPage: number
    nextPage: number
    page: number
    previousPage: number
    total: number
  }
  search: string | null
  sorter: Array<{ column: string; direction: string }>
}
// ...
export interface ProcessApiFilter {
  search: string | null
  filter: Array<Array<{ column: string; operator: string; value: unknown[] }>>
  sorter: Array<{ column: string; direction: string }>
  paging: {
    itemsPerPage: number
    page: number
  }
}